Pushes test code coverage as high as possible. Waiting to finalize inference settings as further changes are needed there for full viability which may result in further code changes.

+++ src/meds_testing_helpers/infer_dataset_config.py
@@ -222,7 +222,7 @@ def main(cfg: DictConfig):
 
     dataset_generator = MEDSDatasetGenerator(
         data_generator=data_generator,
-        shard_size=float(np.median(shard_sizes)),
+        shard_size=int(np.median(shard_sizes)),
         train_frac=train_frac,
         tuning_frac=tuning_frac,
         held_out_frac=held_out_frac,
